Move stellar utils from index to its own folder
